Move NetworkServiceSpy to its own file
