+++ B&WTests/Domain/DefaultLoadImageDataUseCaseTests.swift
@@ -113,44 +113,4 @@ final class DefaultLoadImageDataUseCaseTests: XCTestCase {
         action()
         wait(for: [exp], timeout: 1)
     }
-    
-    private final class NetworkServiceSpy: NetworkService {
-        struct Request {
-            let endpoint: Requestable
-            let completion: CompletionHandler
-        }
-        
-        private var requests = [Request]()
-        var requestCallCount: Int {
-            requests.count
-        }
-        var endpoints: [Requestable] {
-            requests.map(\.endpoint)
-        }
-        
-        struct Cancellable: NetworkCancellable {
-            let afterCancel: () -> Void
-            
-            func cancel() {
-                afterCancel()
-            }
-        }
-        
-        private(set) var cancelCallCount = 0
-        
-        func request(endpoint: Requestable, completion: @escaping CompletionHandler) -> NetworkCancellable? {
-            requests.append(Request(endpoint: endpoint, completion: completion))
-            return Cancellable(afterCancel: { [weak self] in
-                self?.cancelCallCount += 1
-            })
-        }
-        
-        func complete(with error: NetworkError, at index: Int = 0) {
-            requests[index].completion(.failure(error))
-        }
-        
-        func complete(with data: Data?, at index: Int = 0) {
-            requests[index].completion(.success(data))
-        }
-    }
 }
